move `MessageGeneratorContext` to `queue.py`, remove reuse, & auto-close
